further fixes

+++ infobip-spring-data-r2dbc-querydsl/src/main/java/com/infobip/spring/data/r2dbc/QuerydslR2dbcFragment.java
@@ -1,5 +1,7 @@
 package com.infobip.spring.data.r2dbc;
 
+import java.util.function.Function;
+
 import com.querydsl.core.types.Expression;
 import com.querydsl.core.types.Predicate;
 import com.querydsl.sql.SQLQuery;
@@ -7,13 +9,11 @@ import com.querydsl.sql.dml.SQLUpdateClause;
 import org.springframework.r2dbc.core.RowsFetchSpec;
 import reactor.core.publisher.Mono;
 
-import java.util.function.Function;
-
 public interface QuerydslR2dbcFragment<T> {
 
     <O> RowsFetchSpec<O> query(Function<SQLQuery<?>, SQLQuery<O>> builder);
 
-    Mono<Integer> update(Function<SQLUpdateClause, SQLUpdateClause> update);
+    Mono<Long> update(Function<SQLUpdateClause, SQLUpdateClause> update);
 
     /**
      * Deletes all entities matching the given {@link Predicate}.
@@ -21,7 +21,7 @@ public interface QuerydslR2dbcFragment<T> {
      * @param predicate to match
      * @return amount of affected rows
      */
-    Mono<Integer> deleteWhere(Predicate predicate);
+    Mono<Long> deleteWhere(Predicate predicate);
 
     /**
      * Returns entity projection used for mapping {@code QT} to {@code T}.
